% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patients-package.R
\docType{package}
\name{patients-package}
\alias{patients-package}
\alias{_PACKAGE}
\alias{patients}
\alias{package-patients}
\title{patients: Generate adult or pediatric patients with realistic
    characteristics and comorbidities}
\description{
Quickly generate large numbers of rows of patient
    data, with realistic comorbidities and characteristics.
}
\author{
Jack O. Wasey \email{jack@jackwasey.com}
}
\keyword{datagen}
\keyword{internal}
